+++ src/ConnectedMode/Migration/IUnintrusiveBindingController.cs
@@ -29,7 +29,7 @@ namespace SonarLint.VisualStudio.ConnectedMode.Migration
 {
     internal interface IUnintrusiveBindingController
     {
-        Task BindAsync(BoundSonarQubeProject project, IProgress<MigrationProgress> progress, CancellationToken token);
+        Task BindAsync(BoundSonarQubeProject project, IProgress<FixedStepsProgress> progress, CancellationToken token);
     }
 
     [Export(typeof(IUnintrusiveBindingController))]
@@ -44,13 +44,11 @@ namespace SonarLint.VisualStudio.ConnectedMode.Migration
             this.bindingProcessFactory = bindingProcessFactory;
         }
 
-        public async Task BindAsync(BoundSonarQubeProject project, IProgress<MigrationProgress> progress, CancellationToken token)
+        public async Task BindAsync(BoundSonarQubeProject project, IProgress<FixedStepsProgress> progress, CancellationToken token)
         {
-            var progressConverter = new FixedStepsProgressToMigrationProgressConverter(progress);
-
             var bindingProcess = CreateBindingProcess(project);
 
-            await bindingProcess.DownloadQualityProfileAsync(progressConverter, token);
+            await bindingProcess.DownloadQualityProfileAsync(progress, token);
             bindingProcess.SaveRuleConfiguration(token);
             await bindingProcess.SaveServerExclusionsAsync(token);
         }
